+++ src/js/handler/modal/scroll/onModalScroll.js
@@ -1,10 +1,14 @@
-import { YOUTUBE, LOCAL_STORAGE_KEY, BASE_URL } from '../../utils/constant.js';
-import { request } from '../../utils/fetch.js';
-import storage from '../../utils/localStorage.js';
-import { hideElement, showElement } from '../../utils/setAttribute.js';
-import { renderExtraClips } from '../../view/modal.js';
-import { getQueryString } from '../../utils/getQueryString.js';
-import $DOM from '../../utils/DOM.js';
+import {
+  YOUTUBE,
+  LOCAL_STORAGE_KEY,
+  BASE_URL,
+} from '../../../utils/constant.js';
+import { request } from '../../../utils/fetch.js';
+import storage from '../../../utils/localStorage.js';
+import { hideElement, showElement } from '../../../utils/setAttribute.js';
+import { renderExtraClips } from '../../../view/modal.js';
+import { getQueryString } from '../../../utils/getQueryString.js';
+import $DOM from '../../../utils/DOM.js';
 
 export const onModalScroll = async (event) => {
   const { scrollTop, scrollHeight, clientHeight } = event.target;
