+++ src/js/index.js
@@ -1,20 +1,18 @@
 import storage from './utils/localStorage.js';
 import { throttling } from './utils/throttling.js';
 import { LOCAL_STORAGE_KEY } from './utils/constant.js';
-import { onModalShow } from './handler/modal/onModalShow.js';
-import {
-  onSearchByKeyword,
-  onSearchClip,
-} from './handler/modal/onSearchClip.js';
-import { onModalScroll } from './handler/modal/onModalScroll.js';
-import { onSaveClip } from './handler/modal/onSaveClip.js';
-import { onModalClose } from './handler/modal/onModalClose.js';
+import { onModalShow } from './handler/modal/visibility/onModalShow.js';
+import { onModalScroll } from './handler/modal/scroll/onModalScroll.js';
+import { onSaveClip } from './handler/modal/save/onSaveClip.js';
+import { onModalClose } from './handler/modal/visibility/onModalClose.js';
 import { renderSavedClips } from './view/main.js';
 import { hideElement, showElement } from './utils/setAttribute.js';
 import { onToggleRenderedClips } from './handler/main/onToggleRenderedClips.js';
 import { onButtonContainer } from './handler/main/onButtonContainer.js';
 import { onWindowInput } from './handler/window/onWindowInput.js';
 import $DOM from './utils/DOM.js';
+import { onSearchClip } from './handler/modal/search/onSearchClip.js';
+import { onSearchByKeyword } from './handler/modal/search/onSearchBykeyword.js';
 
 const clearDeletedClip = () => {
   const savedClips = storage.get(LOCAL_STORAGE_KEY.SAVED_CLIPS) ?? [];
